chore(rename): change all references to pep8 to pycodestyle
Updates all usages of pep8 to pycodestyle. This is important
a) pep8 no longer exists on the PyPackage Index
   (it has been replaced by pycodestyle)
b) community consensus is that tooling should not act in the name
   of a PEP style guide

For context, PEP8 is a style guide only, and there should not
be any tools in Guido's opinion that bosses people around in its
name. Any issues with the tool is a tool issue, not a PEP issue,
and renaming the pep8 tool away from the PEP makes this clear.