fix(base_model): change datetime usage from utc to unspecified
test_datetime_attributes was failing because it used the system
timezone whereas the BaseModel uses the utc timezone.

Instead of changing the test to match the BaseModel, I have changed
the BaseModel to match the test. In my opinion, it is more useful
to have BaseModel update and creation times expressed in terms of
our development team's system times, as we are currently in the
same timezone and such times are easier to interpret.

If the development team is ever not in the same timezone, this change
may become a problem.